feat: 药企进销存与采购确认入库权限
采购入库确认改为独立权限 inport:confirm，由管理端授给保管员；员工登录后才能过账，未验收合格不能入库存。

+++ warehouse-back/src/main/java/com/sunlee/bus/service/ICustomerService.java
@@ -1,21 +1,9 @@
 package com.sunlee.bus.service;
 
-import com.sunlee.bus.entity.Customer;
 import com.baomidou.mybatisplus.extension.service.IService;
+import com.sunlee.bus.entity.Customer;
 
-/**
- * <p>
- * InnoDB free: 9216 kB 服务类
- * </p>
- *
- * @author sunlee
- * @since 2026-03-15
- */
 public interface ICustomerService extends IService<Customer> {
 
-    /**
-     * 根据客户id删除客户
-     * @param id    客户id
-     */
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
 }
